+++ sqlmesh_dag_generator/__init__.py
@@ -2,7 +2,7 @@
 SQLMesh DAG Generator - Open Source Airflow Integration for SQLMesh
 """
 
-__version__ = "0.9.8"
+__version__ = "0.9.9"
 
 
 from sqlmesh_dag_generator.generator import SQLMeshDAGGenerator